# stats: use sync.Map instead of map + mutex