A [`sync.Map`](https://golang.org/pkg/sync/#Map) is better suited for this workload - so use it.  Additionally, change the `NewPerInstance*` methods to not allocate when the tags argument is empty.

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkParallelCounter-8     55.5          16.8          -69.73%
```